\name{NetworkComparisonTest-package}
\alias{NetworkComparisonTest-package}
\docType{package}
\title{
Statistical comparison of two networks based on the difference in strength
}
\description{
This permutation based hypothesis test, suited for gaussian and binary data, assesses the difference in strength of two networks. Network structures are estimated with l1-regularized partial correlations (gaussian data) or with l1-regularized logistic regression (eLasso, binary data).
}
\details{
\tabular{ll}{
Package: \tab NetworkComparisonTest\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-03-01\cr
License: \tab GPL-2\cr
}
}
\author{
Claudia D. van Borkulo

Maintainer: Claudia D. van Borkulo <cvborkulo@gmail.com>
}
\references{
van Borkulo, C. D., Waldorp, L. J., Boschloo, L., Schoevers, R.A., & Borsboom, D. (2015). Distinguishing networks: A permutation test for comparing network structures. Manuscript in preparation. 
}
% ~~ Optionally other standard keywords, one per line, from ~~
% ~~ file KEYWORDS in the R documentation directory ~~

