% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3Methods.R
\name{summary.NCT}
\alias{summary.NCT}
\title{Summary method for NCT}
\usage{
\method{summary}{NCT}(object, ...)
}
\arguments{
\item{object}{output of NCT}

\item{...}{for now ignored}
}
\description{
Summary method for NCT
}
