\name{NetworkComparisonTest-package}
\alias{NetworkComparisonTest-package}
\docType{package}
\title{
Statistical comparison of two networks based on the difference in strength
}
\description{
This permutation based hypothesis test, suited for gaussian and binary data, assesses the difference in strength of two networks. Network structures are estimated with l1-regularized partial correlations (gaussian data) or with l1-regularized logistic regression (eLasso, binary data). Suited for comparison of independent and dependent samples.
}
\details{
\tabular{ll}{
Package: \tab NetworkComparisonTest\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2015-10-07\cr
License: \tab GPL-2\cr
}
}
\author{
Claudia D. van Borkulo, with contributions from Sacha Epskamp and Alex Millner

Maintainer: Claudia D. van Borkulo <cvborkulo@gmail.com>
}
\references{
Ernst MD. Permutation methods: A basis for exact inference. Stat Sci. 2004;19(4):676-685.

Good PI. Permutation, parametric and bootstrap tests of hypotheses. Vol. 3. New York:: Springer, 2005.

van Borkulo, C. D., Boschloo, L., Borsboom, D., Penninx, B. W. J. H., Waldorp, L. J., & Schoevers, R.A. (2015). Association of symptom network structure with the course of depression. In press.

van Borkulo, C. D., Waldorp, L. J., Boschloo, L., Schoevers, R.A., & Borsboom, D. (2015). Distinguishing networks: A permutation test for comparing network structures. Manuscript in preparation. 
}
% ~~ Optionally other standard keywords, one per line, from ~~
% ~~ file KEYWORDS in the R documentation directory ~~

