\name{NetworkComparisonTest-package}
\alias{NetworkComparisonTest-package}
\docType{package}
\title{
Statistical Comparison of Two Networks Based on Three Invariance Measures
}
\description{
This permutation based hypothesis test, suited for gaussian and binary data, assesses the difference between two networks based on several invariance measures (network structure invariance, global strength invariance, edge invariance). Network structures are estimated with l1-regularized partial correlations (gaussian data) or with l1-regularized logistic regression (eLasso, binary data). Suited for comparison of independent and dependent samples (currently, only for one group measured twice).
}
\details{
\tabular{ll}{
Package: \tab NetworkComparisonTest\cr
Type: \tab Package\cr
Version: \tab 2.1.1\cr
Date: \tab 2018-05-12\cr
License: \tab GPL-2\cr
}
}
\author{
Claudia D. van Borkulo, with contributions from Jonas Haslbeck, Sacha Epskamp and Alex Millner

Maintainer: Claudia D. van Borkulo <cvborkulo@gmail.com>
}
\references{
Ernst, M.D. Permutation methods: A basis for exact inference. Stat Sci. 2004;19(4):676-685.

Good, P.I. Permutation, parametric and bootstrap tests of hypotheses. Vol. 3. New York:: Springer, 2005.

van Borkulo, C. D., Boschloo, L., Borsboom, D., Penninx, B. W. J. H., Waldorp, L. J., & Schoevers, R.A. (2015). Association of symptom network structure with the course of depression. JAMA Psychiatry. 2015;72(12). doi:10.1001/jamapsychiatry.2015.2079

van Borkulo, C. D., Boschloo, Kossakowski, J., Tio, P., L., Schoevers, R.A., Borsboom, D., & , Waldorp, L. J. (2016). Comparing network structures on three aspects: A permutation test. Manuscript submitted. 
}
% ~~ Optionally other standard keywords, one per line, from ~~
% ~~ file KEYWORDS in the R documentation directory ~~

